moved denom_validation to denom file in the common folder for easier implementation for other contracts
