+++ contracts/non-fungible-tokens/andromeda-auction/src/contract.rs
@@ -14,6 +14,7 @@ use andromeda_std::{
     amp::AndrAddr,
     common::{
         actions::call_action,
+        denom::{validate_denom, SEND_CW20_ACTION},
         encode_binary,
         expiration::{expiration_from_milliseconds, get_and_validate_start_time},
         rates::get_tax_amount,
@@ -39,7 +40,6 @@ const CONTRACT_NAME: &str = "crates.io:andromeda-auction";
 const CONTRACT_VERSION: &str = env!("CARGO_PKG_VERSION");
 
 const SEND_NFT_ACTION: &str = "SEND_NFT";
-const SEND_CW20_ACTION: &str = "SEND_CW20";
 
 #[cfg_attr(not(feature = "library"), entry_point)]
 pub fn instantiate(
@@ -253,21 +253,6 @@ pub fn handle_receive_cw20(
             &sender,
         ),
     }
-}
-
-fn validate_denom(deps: DepsMut, env: Env, denom: String) -> Result<(), ContractError> {
-    let potential_supply = deps.querier.query_supply(denom.clone())?;
-    let non_empty_denom = !denom.is_empty();
-    let non_zero_supply = !potential_supply.amount.is_zero();
-    let valid_cw20 = ADOContract::default()
-        .is_permissioned_strict(deps.storage, env, SEND_CW20_ACTION, denom.clone())
-        .is_ok();
-    ensure!(
-        (non_empty_denom && non_zero_supply) || valid_cw20,
-        ContractError::InvalidAsset { asset: denom }
-    );
-
-    Ok(())
 }
 
 #[allow(clippy::too_many_arguments)]
